+++ frontend/src/hooks/list.hooks/useListFilter.ts
@@ -1,9 +1,9 @@
+import { numTermsToReviewState } from "components/review/state/reviewSelectors";
 import { useQueryListsById } from "gql/hooks/list-query";
 import { filterTermsBySaturation } from "helpers/filterTermsBySaturation";
 import { useRouteProps } from "hooks/routerHooks";
 import { useMemo, useState } from "react";
 import { useRecoilValue } from "recoil";
-import { numTermsToReviewState } from "state/selectors/reviewSelectors";
 import { FilterInterface, TruncatedTerm } from "types/list.types";
 
 export function useListFilter() {
